--- backend/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "Back end for Spektate dashboard",
-  "main": "build/index.js",
-  "repository": "https://github.com/microsoft/spektate",
-  "license": "MIT",
-  "private": false,
-  "scripts": {
-    "build": "tsc",
-    "start": "ts-node src/index.ts"
-  },
-  "dependencies": {
-    "@types/express": "^4.17.0",
-    "body-parser": "^1.19.0",
-    "spektate": "file:../packages/spektate",
-    "tslib": "^1.10.0"
-  }
-}
