--- backend/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "outDir": "build",
-    "module": "commonjs",
-    "target": "es6",
-    "lib": ["es2017", "es2015", "dom", "es6"],
-    "sourceMap": true,
-    "moduleResolution": "node",
-    "rootDir": "src",
-    "forceConsistentCasingInFileNames": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noImplicitAny": true,
-    "importHelpers": true,
-    "strictNullChecks": true,
-    "suppressImplicitAnyIndexErrors": true,
-    "noUnusedLocals": true
-  },
-  "exclude": ["node_modules", "build", "scripts", "acceptance-tests", "webpack"]
-}
