+++ package.json
@@ -13,11 +13,13 @@
     "react": "^16.8.5",
     "react-dom": "^16.8.5",
     "react-scripts-ts": "^4.0.8",
-    "spektate": "file:./packages/spektate"
+    "spektate": "file:./packages/spektate",
+    "body-parser": "^1.19.0"
   },
   "scripts": {
+    "start-backend": "ts-node src/backend/index.ts",
     "start": "react-scripts-ts start",
-    "build": "react-scripts-ts build",
+    "build": "react-scripts-ts build && tsc src/backend/** --outDir build/backend/",
     "eject": "react-scripts-ts eject",
     "lint": "tslint -p tsconfig.json -c tslint.json src/**/*.ts",
     "refresh": "rm -rf ./node_modules ./package-lock.json && npm install"
