+++ src/client/assets/pipeline-proto.png
