+++ src/client/config.ts
@@ -15,7 +15,9 @@ export let config: { [id: string]: string | undefined } = {
   BACKEND_URL:
     browser.ENV && browser.ENV.REACT_APP_BACKEND_URL !== undefined
       ? browser.ENV.REACT_APP_BACKEND_URL
-      : process.env.REACT_APP_BACKEND_URL,
+      : process.env.REACT_APP_BACKEND_URL
+      ? process.env.REACT_APP_BACKEND_URL
+      : "",
   GITHUB_MANIFEST_USERNAME:
     browser.ENV && browser.ENV.REACT_APP_GITHUB_MANIFEST_USERNAME !== undefined
       ? browser.ENV.REACT_APP_GITHUB_MANIFEST_USERNAME
