+++ src/index.tsx
@@ -1,9 +1,9 @@
 import { initializeIcons } from "office-ui-fabric-react/lib/Icons";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import "./css/dashboard.css";
-import Dashboard from "./Dashboard";
-import registerServiceWorker from "./registerServiceWorker";
+import "./client/css/dashboard.css";
+import Dashboard from "./client/Dashboard";
+import registerServiceWorker from "./client/registerServiceWorker";
 
 initializeIcons();
 ReactDOM.render(<Dashboard />, document.getElementById("root") as HTMLElement);
